Fix: Split test suite
