+++ Makefile
@@ -32,8 +32,9 @@ static-code-analysis-baseline: vendor ## Generates a baseline for static code an
 	.phive/psalm --config=psalm.xml --set-baseline=psalm-baseline.xml
 
 .PHONY: tests
-tests: vendor ## Runs tests with phpunit/phpunit
-	vendor/bin/phpunit
+tests: vendor ## Runs unit and functional tests with phpunit/phpunit
+	vendor/bin/phpunit --testsuite=unit
+	vendor/bin/phpunit --testsuite=functional
 
 vendor: composer.json
 	composer install --no-progress
