+++ phpunit.xml.dist
@@ -19,8 +19,11 @@
         <env name="SYMFONY_DEPRECATIONS_HELPER" value="max[self]=0"/>
     </php>
     <testsuites>
-        <testsuite name="Project Test Suite">
-            <directory>tests</directory>
+        <testsuite name="functional">
+            <directory>tests/Functional/</directory>
+        </testsuite>
+        <testsuite name="unit">
+            <directory>tests/Unit/</directory>
         </testsuite>
     </testsuites>
 </phpunit>
